+++ cgogn/core/ui_modules/mesh_data.h
@@ -45,14 +45,14 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+
 template <typename MESH>
 struct MeshData
 {
 	template <typename T>
 	using Attribute = typename mesh_traits<MESH>::template Attribute<T>;
 	using AttributeGen = typename mesh_traits<MESH>::AttributeGen;
-
-	using Vec3 = geometry::Vec3;
 
 	MeshData() : mesh_(nullptr), bb_vertex_position_(nullptr), bb_min_(0, 0, 0), bb_max_(0, 0, 0), outlined_until_(0.0)
 	{
