+++ cgogn/geometry/algos/medial_axis.h
@@ -41,8 +41,8 @@ namespace cgogn
 namespace geometry
 {
 
-using Vec3 = geometry::Vec3;
-using Scalar = geometry::Scalar;
+using geometry::Vec3;
+using geometry::Scalar;
 
 inline Scalar compute_radius(const Vec3& p, const Vec3& n, const Vec3& q)
 {
