+++ cgogn/geometry/apps/filtering.cpp
@@ -43,8 +43,8 @@ template <typename T>
 using Attribute = typename cgogn::mesh_traits<Mesh>::Attribute<T>;
 using Vertex = typename cgogn::mesh_traits<Mesh>::Vertex;
 
-using Vec3 = cgogn::geometry::Vec3;
-using Scalar = cgogn::geometry::Scalar;
+using cgogn::geometry::Vec3;
+using cgogn::geometry::Scalar;
 
 int main(int argc, char** argv)
 {
