+++ cgogn/geometry/apps/registration.cpp
@@ -41,8 +41,8 @@ using Attribute = typename cgogn::mesh_traits<Mesh>::Attribute<T>;
 using Vertex = typename cgogn::mesh_traits<Mesh>::Vertex;
 using Face = typename cgogn::mesh_traits<Mesh>::Face;
 
-using Vec3 = cgogn::geometry::Vec3;
-using Scalar = cgogn::geometry::Scalar;
+using cgogn::geometry::Vec3;
+using cgogn::geometry::Scalar;
 
 int main(int argc, char** argv)
 {
