+++ cgogn/geometry/ui_modules/surface_differential_properties.h
@@ -43,6 +43,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class SurfaceDifferentialProperties : public Module
 {
@@ -54,9 +57,6 @@ class SurfaceDifferentialProperties : public Module
 
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 public:
 	SurfaceDifferentialProperties(const App& app)
