+++ cgogn/geometry/ui_modules/surface_filtering.h
@@ -40,6 +40,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class SurfaceFiltering : public Module
 {
@@ -50,9 +53,6 @@ class SurfaceFiltering : public Module
 
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 public:
 	SurfaceFiltering(const App& app)
