+++ cgogn/geometry/ui_modules/volume_selection.h
@@ -52,6 +52,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class VolumeSelection : public ViewModule
 {
@@ -74,9 +77,6 @@ class VolumeSelection : public ViewModule
 	{
 		SingleCell
 	};
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 	struct Parameters
 	{
