+++ cgogn/io/incidence_graph/incidence_graph_import.h
@@ -39,7 +39,7 @@ namespace cgogn
 namespace io
 {
 
-using Vec3 = geometry::Vec3;
+using geometry::Vec3;
 
 struct IncidenceGraphImportData
 {
