+++ cgogn/modeling/algos/decimation/QEM_helper.h
@@ -26,15 +26,21 @@
 
 
 #include <cgogn/core/functions/traversals/vertex.h>
-
+#include <cgogn/core/types/mesh_traits.h>
+#include <cgogn/core/utils/type_traits.h>
 #include <cgogn/geometry/types/quadric.h>
 #include <cgogn/geometry/types/vector_traits.h>
 
 namespace cgogn
 {
 
+
 namespace modeling
 {
+
+using geometry::Vec3;
+using geometry::Scalar;
+using geometry::Quadric;
 
 template <typename MESH>
 struct DecimationQEM_Helper
@@ -45,10 +51,6 @@ struct DecimationQEM_Helper
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
 	using Face = typename mesh_traits<MESH>::Face;
-
-	using Vec3 = typename geometry::Vec3;
-	using Scalar = typename geometry::Scalar;
-	using Quadric = typename geometry::Quadric;
 
 	DecimationQEM_Helper(MESH& m, const Attribute<Vec3>* vertex_position) : m_(m), vertex_position_(vertex_position)
 	{
