+++ cgogn/modeling/algos/decimation/decimation.h
@@ -36,8 +36,8 @@ namespace cgogn
 namespace modeling
 {
 
-using Vec3 = geometry::Vec3;
-using Scalar = geometry::Scalar;
+using geometry::Vec3;
+using geometry::Scalar;
 
 /////////////
 // GENERIC //
