+++ cgogn/modeling/algos/graph_to_hex.h
@@ -41,9 +41,9 @@ class CellMarker;
 namespace modeling
 {
 
-using Vec3 = geometry::Vec3;
+using geometry::Vec3;
 using Vec3i = geometry::Vec3i;
-using Scalar = geometry::Scalar;
+using geometry::Scalar;
 using Mat3 = geometry::Mat3;
 
 struct GAttributes
