+++ cgogn/modeling/algos/subdivision.h
@@ -37,7 +37,7 @@ namespace cgogn
 namespace modeling
 {
 
-using Vec3 = geometry::Vec3;
+using geometry::Vec3;
 
 ///////////
 // CMap3 //
