+++ cgogn/modeling/algos/subdivision/surface_loop.h
@@ -33,8 +33,8 @@ namespace cgogn
 namespace modeling
 {
 
-using Vec3 = geometry::Vec3;
-using Scalar = geometry::Scalar;
+using geometry::Vec3;
+using geometry::Scalar;
 
 ///////////
 // CMap2 //
