+++ cgogn/modeling/algos/volume_utils.h
@@ -33,7 +33,7 @@ namespace cgogn
 namespace modeling
 {
 
-using Vec3 = geometry::Vec3;
+using geometry::Vec3;
 
 //////////
 // CMap //
