+++ cgogn/modeling/ui_modules/surface_deformation.h
@@ -47,6 +47,9 @@ namespace cgogn
 
 namespace ui
 {
+using geometry::Vec3;
+using geometry::Mat3;
+using geometry::Scalar;
 
 template <typename MESH>
 class SurfaceDeformation : public ViewModule
@@ -58,10 +61,6 @@ class SurfaceDeformation : public ViewModule
 
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
-
-	using Vec3 = geometry::Vec3;
-	using Mat3 = geometry::Mat3;
-	using Scalar = geometry::Scalar;
 
 	struct Parameters
 	{
