+++ cgogn/modeling/ui_modules/surface_modeling.h
@@ -48,6 +48,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class SurfaceModeling : public Module
 {
@@ -59,9 +62,6 @@ class SurfaceModeling : public Module
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
 	using Face = typename mesh_traits<MESH>::Face;
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 public:
 	SurfaceModeling(const App& app)
