+++ cgogn/modeling/ui_modules/volume_deformation.h
@@ -43,6 +43,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class VolumeDeformation : public ViewModule
 {
@@ -53,9 +56,6 @@ class VolumeDeformation : public ViewModule
 
 	using Vertex = typename mesh_traits<MESH>::Vertex;
 	using Edge = typename mesh_traits<MESH>::Edge;
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 	struct Parameters
 	{
