+++ cgogn/rendering/ui_modules/vector_per_vertex_render.h
@@ -44,6 +44,9 @@ namespace cgogn
 namespace ui
 {
 
+using geometry::Vec3;
+using geometry::Scalar;
+
 template <typename MESH>
 class VectorPerVertexRender : public ViewModule
 {
@@ -51,9 +54,6 @@ class VectorPerVertexRender : public ViewModule
 	using Attribute = typename mesh_traits<MESH>::template Attribute<T>;
 
 	using Vertex = typename mesh_traits<MESH>::Vertex;
-
-	using Vec3 = geometry::Vec3;
-	using Scalar = geometry::Scalar;
 
 	struct Parameters
 	{
